Extra cellular bacteris are no longer die when a compartment becomes caseated. They become non-replicating bacteria.
There are 2 new statistics, the number of caseated compartments and the number of non-replicating bacteria. These are written to the statistics file and displayed on the screen in both the GUI and non-GUI version.

Changed the probability of a chronically infected macrophage uptaking bacteria to be 1/2 the probability that a resting macrophage uptakes bacteria.

Made the statistics file written by the non-GUI version consistent with the one written by the GUI version.

git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@718 88ce85bb-8564-4ae6-884f-ab7503230054