Fixed label for non-replicating bacteria in displays and in stats file header.
Fixed statswidget in GUI - delete an empty row, and resized so the bottom is no longer cutoff.

GridCell no returns the number of Be even for caseated comaprtments.

If a compartment becomes caseated all immune cells are killed, including macrophages, not just T cells. GridCell::getExtMtb returns the number of bacteria in a compartment - it no longer returns 0 if the compartment is caseated.

git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@722 88ce85bb-8564-4ae6-884f-ab7503230054