Added some extra information to saved states to (sometimes) catch serialization/deserialization errors.
* The grid dimensions are written and checked on load of a saved state.

* The size of each object (from sizeof) is written and checked on load of a saved state. This is not foolproof because of data alignment. An object can have a data member added or deleted and sizeof may report the same size.

Changed the order of serialization of data members to exactly match their sequence in the class declaration, allowing easier checking for correctness.

Fixed bugs in save and load of simulation state:

* Lock the simulation object before taking snapshots so the simulation doesn't advance while saving state.

* Added default constructors for the agent classes and use them when creating agents during load of a saved state, avoiding use of the random number generator.

However, running a simulation from a load of a saved state still does not produce the exactly the same results as the initial run that produced the saved state.

In the Rand class include <boost/tr1/random.hpp> rather than <boost/tr1> for Linux builds. For reasons unknown the random number generator won't synch properly when deserialized.

In the GUI version snapshots are taken on the last time step even if that isn't one for a snapshot interval.


git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@724 88ce85bb-8564-4ae6-884f-ab7503230054