Doesn't write a statistics file header when no statistics output requested (--csv-interval is 0).
Fixed a bug where the simulation limit was not properly defined if for a limit of 0 timesteps. This causes a problem when using the GUI version to load a state and save a snapshot for a state for time step 0.

Changed the default simulation limit for the Simulation class to be a value in time steps rather than days, since it is now treated as a number of time steps to simulate, not days.

git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@736 88ce85bb-8564-4ae6-884f-ab7503230054