Changed lhs program to allow parameters with a single value instead of a range. Ranges now only need to be specified for parameters that are being varied. This makes LHS parameter files easier to read.
Added a function, ParamsBase::paramIndex, to convert from an index into the integer parameter value array to an index into the parameter description. This replaces hard coded conversions.

Fixed a bug in the Lhs::checkParams function. The logical expression for choosing between checking a double parameter and checking an integer parameter wasn't correct.

git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@767 88ce85bb-8564-4ae6-884f-ab7503230054