Deleted writing/reading of object size when serializing/deserializing objects, because it is not portable across 32 bit and 64 bit systems - a state saved on a 64 bit system cannot be loaded on a 32 bit system, and vice/versa.
Added a text header and footer for each object serialized, which is read and checked when deserializing. This protects against some serialization errors and is portable across 32 bit and 64 bit systems.

Fixed a bug in the definition of _NFkB_IkB the Macrophage constructor. If parameter meanNFkB is 0 then _NFkB_IkB would be set to nan (not a number) for release builds and cause a run-time assertion error in the Boost library lognormal_distribution constructor. The fix is to set it to 0 in that case.


git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@773 88ce85bb-8564-4ae6-884f-ab7503230054