Fixed a bug where seeds loaded from a saved state where truncated in the gui version. THis was due to a hardcoded limit of 10000000 in the mainwindow.ui file. This is now set on program initialization to a limit of INT_MAX (since the QSpinBox widget for the seed handles values of type int).
Changed all variables that process seed values to be of type unsigned int rather than unsigned long since the Rand class treats seeds as unsigned int and time values used to define seeds are of that type also.

Created function createTimeSeed to define a seed based on time, so that is done consistently between the gui and non-gui versions.

git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@1273 88ce85bb-8564-4ae6-884f-ab7503230054