Changed Treg deactivation to not deactivate any immune cells that have a state or next state of dead. This can resurrect cells that previously died on the current time step. If the resurrected cell is in a caseated compartment this can leave the simulation in an inconsistent state if the simulation state is saved and loaded. On state load the resurrected cell will be added to a cell list, but won't be added to its grid compartment because cells cannot be added to a caseated compartment.
Changed the Treg deactivation so it attempts to deactivate all immune cells in a grid compartment. Previously if a grid compartment had more than one cell, if the first cell was deactivated no attempt would be made to deactivate the other cell in the compartment.

Changed PARAM_TREG_PROB_DOWN_REGULATE to PARAM_TREG_PROB_DEACTIVATE to be consistent with the 3D lung model.

git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@1356 88ce85bb-8564-4ae6-884f-ab7503230054