Added stat DrugConcentrationAreaINH for calculating the drug area under the curve for INH.
Added parameter drugConcentrationStatInterval for specifying the number of timesteps to
use for stat DrugConcentrationArea.

Changed so drug dynamics values are always written to the stats file even if drug dynamics not enabled,
so stats file always has the same format. This is useful for post processing of the stats file.



git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@1867 88ce85bb-8564-4ae6-884f-ab7503230054